Add files via upload
This script defines a list of strings and a function is_long() that returns True if a string is longer than 3 characters. It then uses the filter() function to filter the list of strings, keeping only the strings that are longer than 3 characters. The resulting list is then printed.

In this example, the output will be: ['hello', 'world']. This is because the is_long() function returns True for the strings 'hello' and 'world', but False for the other strings in the list.